AMP-14441
each filter load is, since 2.4, done in two stages:
a) fill filter with default values
b) overwrite filter fields which are stored in the database

I added step:
a.2) store filter's default values (as a String representation) in a variable. 
This is not an AMP-wide constant because the "filter default values" are a function of, among others, user and workspace settings.

Now the warning note is displayed each time the filters have been changed compared to the DEFAULT values (saved in the variable named above).

ALSO: fixed:
 1) translations bug in ColumnFilterData.jsp
 2) in certain situations tab filters were not loaded from the DB

git-svn-id: file:///svn/branches/AMP_2_4_RELEASE@19725 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4