AMP-15765
unify the source of the settings used when deciding whether projects should be validated or no. Now all the places in the code firstly look at the team-member settings and, if those are missing, look at the workspace settings.
use workspace settings (ss opposed to team-member settings) when deciding whether a new activity should be validated. 

These decisions all go through a couple of unified places in the code - in preparation for the upcoming AMP-15891



git-svn-id: file:///svn/branches/AMP_2_4_RELEASE@20962 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4