AMP-15967
preparation for doing the debugging.
Fix the metadata storing "subsystem" in the reports whereas MetaInfo is stored inconsistently either in hashsets or treesets; is treated either as a Comparable (with compareTo comparing both category and value) or as a plain object (hashCode() comparing by category only, equals() comapring both category and value). Also, fixing the storage allowed removing manual iteration of all of a cell's metadata in inner loops of the reports engine.
Now, reports with 2-3-deep hierarchies run about 25% faster just because of this

git-svn-id: file:///svn/branches/AMP_2_6_RELEASE@21511 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4