AMP-17806: Refactor the app's entry and config so it does not depend on html at all
This should make using the dojo build system possible.

The tricky part is using AMD to load the config... before configuring AMD. It is resolved by using an absolute path to the config. DOJO AMD's  optionally updates configuration if an object is passed as a first parameter, so this mechanism is used to set the configuration for the whole app.

Hooking into dojo build should now be a matter of scraping all the required sources, and then running the script.

git-svn-id: file:///svn/branches/AMP_2_10_RELEASE@23822 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4