AMP-18560: Fix saved dashboards
several small bugs

refactoring state into the request required a bigger refactor of the app init sequence, since loading state is now async.

Some date stuff was broken in the filters. I added a hack to make it work, but it needs work possibly on the API end to do dates right.

git-svn-id: file:///svn/branches/AMP_2_10_RELEASE@25746 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4